requirements.txt
webdriver_manager
selenium
aiohttp
beautifulsoup4
colorama
rich
requests
gitpython
prompt_toolkit
pyyaml
Flask
streamlit
pandas
plotly
streamlit-option-menu
streamlit-aggrid
